Add skybox infrastructure, code cleanup, and keybind for bubbling effect
Here, I added the infrastructure for adding a skybox. Currently, the panoramic image is of the Milky Way sky, though this is just a placeholder (we can replace it with an image that makes more sense later). I have also cleaned up some of the code that has become obsolete (mouseX, mouseY), as well as condense some of the code to be more neat (moving the mousemove event listener function into just one function). Finally, I've added a keybind for the bubbling effect (before, you just would hover the mouse on top of the galaxy to cause the effect; now this only happens when you hold 'e').