Change elliptical galaxy implementation
The elliptical galaxy implementation was modified a bit to be more spherical. Moreover, the mixedColors effect was modified so that stars are more yellow at the center and gradually become more blue the further from the galactic center they are. Finally, a density function was used so that the stars are denser closer to the galactic center. Credit: ChatGPT was consulted for determining the positioning of each of the particles to be an elliptical shape. It also helped me determine a suitable star density function.