Code refactoring, invisible plane, and blackhole effect
Here, I have refactored some of the code (e.g. variable names to be more descriptive, moving certain sections, etc). I have also changed the invisible plane implementation so that the plane simply lies in the same plane as the galaxy to ensure consistency. Moreover, I have also added the black hole effect, which will suck particles towards the cursor position while the 'f' key is being held. Next, I would like to change it so that the effects are more consistent; currently, they rotate with the galaxy, which is unintended behavior. I want to change how particles in the galaxies are updated to prevent this behavior.