Elliptical Galaxy Changes and Position Updates
In this update, I have overhauled how particle updates occur. Instead of just rotating the mesh, which can lead to some galaxies looking awkward (elliptical), I instead changed it so that functions for each respective galaxy type are called to handle this. I have also created a "lag effect", in which particles further away from the galactic center have a slower rotation speed.

Known bug: the effects (blackhole and repulsion) don't work for the elliptical galaxy, and this has to do with how the particle updates are designed. I can try to address this, but right now these functions either expand or flatten the galaxy (since the implementation only allows the y-axis to be updated), which is also an interesting effect.

Next on the agenda: implement irregular galaxies and quasars. Find a better skybox. Create a somewhat better UI (perhaps using Figma; not sure yet). Add custom objects for each particle to look more like stars.